Changes to register page and check dynamodb connection

+++ backend/app/routes/auth.py
@@ -20,8 +20,8 @@ async def register_user(user_data: Dict):
             "username": user_data["username"],
             "email": user_data["email"],
             "password_hash": hashed_password,
-            "height": float(user_data["height"]) if user_data.get("height") else None,
-            "weight": float(user_data["weight"]) if user_data.get("weight") else None,
+            "height": int(user_data["height"]) if user_data.get("height") else None,
+            "weight": int(user_data["weight"]) if user_data.get("weight") else None,
             "fitness_level": user_data.get("fitness_level", "beginner"),
             "goals": user_data.get("goals", [])
         }
